LayoutTests/imported/w3c/web-platform-tests/wasm/webapi/esm-integration/invalid-bytecode.tentative-expected.txt

CONSOLE MESSAGE: CompileError: WebAssembly.Module doesn't parse at byte 8: invalid section
CONSOLE MESSAGE: CompileError: WebAssembly.Module doesn't validate: control flow returns with unexpected type. I64 is not a I32, in function at index 0

PASS Test that imports of invalid WebAssembly modules leads to WebAssembly.CompileError on window.
